requirements.txt
django
django-stubs
pillow
python-decouple
flake8
ipdb
requests
django-debug-toolbar
django-silk
faker
pytest-django
pytest-cov
pytest-factoryboy
django-widget-tweaks
django-lifecycle
celery[redis]
django-celery-results
django-redisboard
django-phonenumber-field[phonenumberslite]
pytest-mock
freezegun
beautifulsoup4